Fix parser/preparser validation of empty strings (#2748)
* add validator tests

* add notEmpty test

* add blank string to value_is_empty values

* replace removed text

* fix partial oob string indexing

---------
